Move modules to src
The number of auxilliary directories is growing, and I can do this now since the repository isn't stored directly in Scriptable
