+++ src/example/index.ts
