+++ src/example/module.json
